requirements.txt
pandas
dash
plotly
psycopg2-binary
redis
kafka-python
fastapi
uvicorn
gtfs-realtime-bindings
pytest
streamlit